% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEG.R
\name{estim_mu_sigma_nu_tau_GEG}
\alias{estim_mu_sigma_nu_tau_GEG}
\title{estim_mu_sigma_nu_tau_GEG}
\usage{
estim_mu_sigma_nu_tau_GEG(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for GEG distribution.
}
\examples{
y <- rGEG(n=100, mu=1, sigma=1, nu=1, tau=1)
estim_mu_sigma_nu_tau_GEG(y=y)
}
