% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEG.R
\name{logLik_GEG}
\alias{logLik_GEG}
\title{logLik_GEG}
\usage{
logLik_GEG(logparam = c(0, 0, 0, 0), x)
}
\arguments{
\item{logparam}{vector with the values for mu, sigma, nu and tau}

\item{x}{vector with the data}
}
\description{
This is an auxiliar function to obtain the logLik for GEG.
}
\examples{
y <- rGEG(n=100, mu=1, sigma=1, nu=1, tau=1)
logLik_GEG(logparam=c(0, 0, 0, 0), x=y)
}
